oncosegunda/requirements.txt
# requirements.txt
# Dependencias principales de Django
Django==5.0
# psycopg2-binary==2.9.9  # Para PostgreSQL - instalar manualmente en Windows
redis==5.0.8
django-redis==5.4.0

# Dependencias de UI y frontend
django-bootstrap5==23.4
django-crispy-forms==2.3
crispy-bootstrap5==0.7

# Dependencias de seguridad y autenticación
django-allauth==0.63.6
bcrypt==4.2.0
PyJWT==2.9.0

# Dependencias de archivos y storage
Pillow==10.4.0
boto3==1.35.0
django-storages==1.14.4

# Dependencias de testing
pytest==8.3.0
pytest-django==4.8.0
factory-boy==3.3.0
faker==25.0.0

# Dependencias de desarrollo
black==24.8.0
flake8==7.1.0
isort==5.13.2

# Dependencias de producción
gunicorn==23.0.0
whitenoise==6.7.0

# Dependencias de monitoreo y logging
sentry-sdk==2.10.0

# Dependencias adicionales
django-extensions==3.2.3
django-debug-toolbar==4.4.0
django-compressor==4.4
django-ratelimit==4.1.0
django-health-check==3.18.1
django-cors-headers==4.4.0

# Dependencias de tareas programadas (opcional)
celery==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.6.0